move where runmode is set
