+++ dsl/pace/dsl/stencil.py
@@ -336,21 +336,6 @@ class FrozenStencil(SDFGConvertible):
         # If we orchestrate, move the compilation at call time to make sure
         # disable_codegen do not lead to call to uncompiled stencils, which fails
         # silently
-
-        if stencil_config.compilation_config.run_mode == RunMode.Run:
-
-            def exit_instead_of_build(self):
-                stencil_class = None if self.options.rebuild else self.backend.load()
-                if stencil_class is None:
-                    raise RuntimeError(
-                        "Stencil needs to be compiled first in run mode, exiting"
-                    )
-                return stencil_class
-
-            from gt4py.stencil_builder import StencilBuilder
-
-            StencilBuilder.build = exit_instead_of_build
-
         if self.stencil_config.dace_config.is_dace_orchestrated():
             self.stencil_object = gtscript.lazy_stencil(
                 definition=func,
